0.6.2 (20040102)
- Fixed bug: --stop during spawn delay wasn't acted upon
- Fixed API bug: <slack/lib.h> didn't include <slack/coproc.h>
- Fixed bug: -lm wasn't in `libslack-config --libs` on Solaris
- Changed --restart: reset spawn burst counters (for stodden at peppermind.de)
- Changed daemon_path_is_safe() to give an explanation when unsafe
- Added explanations when rejecting unsafe config files and executables
- Fixed bug: reset agent state to IDLE when interrupted (bte at kamash.com)
- Trim unquoted leading spaces from property values (with bte at kamash.com)
- Also trim only unquoted trailing spaces from property names
- Added tools/migrate-properties utility (preserves old propfile behaviour)
- Fixed bug: ownership partially lost in map_resize (with bte at kamash.com)
- Fixed bug: mem_resize was broken since libslack-0.4 (bte at kamash.com)
- Fixed DOC bug: stated the importance of including <slack/lib.h> first
- Cleaned up error messages (now that prog name is set to --name argument)
- Added sections to libslack(3) features list (coproc+pty, low level api)
- Fixed bug: if select() failed (unlikely), coproc wasn't closed properly
- Fixed bug: when tty_raw() failed (unlikely), it returned the wrong value
- Added many examples to the manpages (there are now 91 runnable examples)
- Added tools/check-examples to verify that all examples work
- Fixed bug: str_fgetline() returned empty string rather than null on eof
- Fixed bug: cstrstr() didn't always work due to a typo
- Added sections to libslack(3) features list (documentation, testing)
- Added intel solaris8 binary package
- Ported to Mac OS X (Darwin) 10.3.2
- Added make osx osx-daemon osx-slack (macosx binary packages)
- Added pid to --running output (requested by rubinson at email.arizona.edu)
- Dropped support for K&R clients (suggested by skaller at ozemail.com.au)
- Added hsort_closure(3) (suggested by skaller at maxtal.com.au)
- Added make slack.swig (generate a SWIG input file for libslack)

+++ libslack/sig.c
@@ -1,7 +1,7 @@
 /*
 * libslack - http://libslack.org/
 *
-* Copyright (C) 1999-2002 raf <raf@raf.org>
+* Copyright (C) 1999-2004 raf <raf@raf.org>
 *
 * This program is free software; you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
@@ -18,7 +18,7 @@
 * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 * or visit http://www.gnu.org/copyleft/gpl.html
 *
-* 20020916 raf <raf@raf.org>
+* 20040102 raf <raf@raf.org>
 */
 
 /*
@@ -458,7 +458,7 @@ static int verify(int test, const char *name, const char *msg1, const char *msg2
 
 	if (strncmp(buf, msg1, msg1_length))
 	{
-		printf("Test %d: msg file produced incorrect input:\nshould be\n%s\nwas:\n%*.*s\n", test, msg1, msg1_length, (int)msg1_length, buf);
+		printf("Test %d: msg file produced incorrect input:\nshould be\n%s\nwas:\n%*.*s\n", test, msg1, (int)msg1_length, (int)msg1_length, buf);
 		return 1;
 	}
